web/backend/requirements.txt
# KissBot Web Backend - Dependencies

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic-settings>=2.0.0

# HTTP client
httpx>=0.25.0

# Templates
jinja2>=3.1.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# CORS
python-multipart>=0.0.6
